D-test/requirements.txt
Flask<3.0
Flask-MySQLdb
python-dotenv 